update examples. demonstrate failing dispatch
